+++ backends/CMakeLists.txt
@@ -96,7 +96,10 @@ message(STATUS "OpenCL_LIBRARY: ${OpenCL_LIBRARY}")
 add_library(
   DPPLSyclInterface
   SHARED
+  source/dppl_sycl_context_interface.cpp
+  source/dppl_sycl_device_interface.cpp
   source/dppl_sycl_queue_interface.cpp
+  source/dppl_sycl_queue_manager.cpp
 )
 
 # Install DPPLOpenCLInterface
