+++ backends/include/error_check_macros.h
@@ -23,6 +23,7 @@
 /// codes.
 ///
 //===----------------------------------------------------------------------===//
+
 #pragma once
 
 #include <stdio.h>
