+++ backends/tests/test_sycl_queue_manager.cpp
@@ -1,6 +1,32 @@
+//===--- test_sycl_queue_manager.cpp - DPPL-SYCL interface --*- C++ ---*---===//
+//
+//               Python Data Parallel Processing Library (PyDPPL)
+//
+// Copyright 2020 Intel Corporation
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+//===----------------------------------------------------------------------===//
+///
+/// \file
+/// This file has unit test cases for functions defined in
+/// dppl_sycl_queue_interface.h.
+///
+//===----------------------------------------------------------------------===//
+#include "dppl_sycl_device_interface.h"
+#include "dppl_sycl_queue_manager.h"
 #include "dppl_sycl_queue_interface.h"
 #include <gtest/gtest.h>
-#include <iostream>
 #include <thread>
 
 using namespace std;
@@ -15,8 +41,8 @@ namespace
         num = DPPLGetNumActivatedQueues();
         DPPLPopSyclQueue();
         DPPLPopSyclQueue();
-        DPPLDeleteQueue(q1);
-        DPPLDeleteQueue(q2);
+        DPPLDeleteSyclQueue(q1);
+        DPPLDeleteSyclQueue(q2);
     }
 
     void bar (size_t & num)
@@ -25,7 +51,7 @@ namespace
         // Capture the number of active queues in second
         num = DPPLGetNumActivatedQueues();
         DPPLPopSyclQueue();
-        DPPLDeleteQueue(q1);
+        DPPLDeleteSyclQueue(q1);
     }
 }
 
@@ -102,7 +128,7 @@ TEST_F (TestDPPLSyclQueuemanager, CheckGetNumActivatedQueues)
     EXPECT_EQ(num2, 1);
     EXPECT_EQ(num4, 0);
 
-    DPPLDeleteQueue(q);
+    DPPLDeleteSyclQueue(q);
 }
 
 
@@ -115,8 +141,8 @@ TEST_F (TestDPPLSyclQueuemanager, CheckDPPLDumpPlatformInfo)
 TEST_F (TestDPPLSyclQueuemanager, CheckDPPLDumpDeviceInfo)
 {
     auto q = DPPLGetCurrentQueue();
-    EXPECT_NO_FATAL_FAILURE(DPPLDumpDeviceInfo(q));
-    EXPECT_NO_FATAL_FAILURE(DPPLDeleteQueue(q));
+    EXPECT_NO_FATAL_FAILURE(DPPLDumpDeviceInfo(DPPLGetDeviceFromQueue(q)));
+    EXPECT_NO_FATAL_FAILURE(DPPLDeleteSyclQueue(q));
 }
