+++ dppl/tests/test_dump_functions.py
@@ -31,6 +31,10 @@ class TestDumpMethods(unittest.TestCase):
     def test_dppl_dump_device_info (self):
         q = dppl.get_current_queue()
         try:
-            dppl.dump_device_info(q)
+            q.get_sycl_device().dump_device_info()
         except Exception:
             self.fail("Encountered an exception inside dump_device_info().")
+
+
+if __name__ == '__main__':
+    unittest.main()
